renamed requests2 to requests and some more work in progress
