+++ scout/commands/download/hpo.py
@@ -4,8 +4,7 @@ import pathlib
 
 import click
 
-from scout.utils.scout_requests2 import (fetch_hpo_terms, fetch_hpo_to_genes_to_disease,
-    fetch_genes_to_hpo_to_disease)
+from scout.utils.scout_requests import (fetch_hpo_terms, fetch_hpo_to_genes_to_disease)
 
 LOG = logging.getLogger(__name__)
 
@@ -21,13 +20,6 @@ def print_hpo(out_dir):
     LOG.info("Download HPO terms to %s", hpo_file_path)
     with hpo_file_path.open("w", encoding="utf-8") as outfile:
         for line in fetch_hpo_terms():
-            outfile.write(line + "\n")
-
-    hpo_file_name = "genes_to_phenotype.txt"
-    hpo_file_path = out_dir / hpo_file_name
-    LOG.info("Download HPO genes to phenotypes to diagnosis data to %s", hpo_file_path)
-    with hpo_file_path.open("w", encoding="utf-8") as outfile:
-        for line in fetch_genes_to_hpo_to_disease():
             outfile.write(line + "\n")
 
     hpo_file_name = "phenotype_to_genes.txt"
