+++ scout/commands/download/omim.py
@@ -4,7 +4,7 @@ import pathlib
 
 import click
 
-from scout.utils.scout_requests2 import fetch_mim_files
+from scout.utils.scout_requests import fetch_mim_files
 
 LOG = logging.getLogger(__name__)
